requirements.txt
flask
python-dotenv
langchain>=0.2.12
langchain-community>=0.2.11
langchain-groq>=0.1.3
chromadb
sentence-transformers
